Final touches to business backend.
Changes what data is stored based on what was in the csv.
Added script in data/ that uploads the csv file.
The script ran in about 15 minutes and uploaded 3099 businesses.
data/failures.txt contains the dat on the 375 rows that failed to get uploaded.
I'm not worrying about them, as 3099 is more than enough for a demo.